// archive/sdk-pre-vnext-20260606/src/runtime/generated/runtime/v1/local_runtime_asset_catalog.ts
export enum LocalAssetKind {
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_KIND_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * Runnable kinds
     *
     * @generated from protobuf enum value: LOCAL_ASSET_KIND_CHAT = 1;
     */
    CHAT = 1,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_KIND_IMAGE = 2;
     */
    IMAGE = 2,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_KIND_VIDEO = 3;
     */
    VIDEO = 3,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_KIND_TTS = 4;
     */
    TTS = 4,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_KIND_STT = 5;
     */
    STT = 5,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_KIND_EMBEDDING = 6;
     */
    EMBEDDING = 6,
    /**
     * Passive kinds
     *
     * @generated from protobuf enum value: LOCAL_ASSET_KIND_VAE = 10;
     */
    VAE = 10,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_KIND_CLIP = 11;
     */
    CLIP = 11,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_KIND_LORA = 12;
     */
    LORA = 12,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_KIND_CONTROLNET = 13;
     */
    CONTROLNET = 13,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_KIND_AUXILIARY = 14;
     */
    AUXILIARY = 14
}
/**
 * @generated from protobuf enum nimi.runtime.v1.LocalAssetStatus
 */
export enum LocalAssetStatus {
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_STATUS_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_STATUS_INSTALLED = 1;
     */
    INSTALLED = 1,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_STATUS_ACTIVE = 2;
     */
    ACTIVE = 2,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_STATUS_UNHEALTHY = 3;
     */
    UNHEALTHY = 3,
    /**
     * @generated from protobuf enum value: LOCAL_ASSET_STATUS_REMOVED = 4;
     */
    REMOVED = 4
}
// ...
export enum LocalProfileEntryKind {
    /**
     * @generated from protobuf enum value: LOCAL_PROFILE_ENTRY_KIND_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: LOCAL_PROFILE_ENTRY_KIND_SERVICE = 3;
     */
    SERVICE = 3,
    /**
     * @generated from protobuf enum value: LOCAL_PROFILE_ENTRY_KIND_NODE = 4;
     */
    NODE = 4,
    /**
     * @generated from protobuf enum value: LOCAL_PROFILE_ENTRY_KIND_ASSET = 5;
     */
    ASSET = 5
}