// archive/sdk-pre-vnext-20260606/src/runtime/generated/runtime/v1/local_runtime_device_environment.ts
export enum GpuMemoryModel {
    /**
     * @generated from protobuf enum value: GPU_MEMORY_MODEL_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: GPU_MEMORY_MODEL_DISCRETE = 1;
     */
    DISCRETE = 1,
    /**
     * @generated from protobuf enum value: GPU_MEMORY_MODEL_UNIFIED = 2;
     */
    UNIFIED = 2
}